move pipeline to stages-api
