+++ crates/stages-api/src/lib.rs
@@ -3,6 +3,7 @@ mod error;
 mod metrics;
 mod pipeline;
 mod stage;
+mod util;
 
 pub use crate::metrics::*;
 pub use error::*;
