+++ crates/stages-api/src/pipeline/builder.rs
@@ -1,8 +1,7 @@
-use crate::{pipeline::BoxedStage, Pipeline, StageSet};
+use crate::{pipeline::BoxedStage, MetricEventsSender, Pipeline, Stage, StageSet};
 use reth_db::database::Database;
 use reth_primitives::{stage::StageId, BlockNumber, B256};
 use reth_provider::ProviderFactory;
-use reth_stages_api::{MetricEventsSender, Stage};
 use reth_static_file::StaticFileProducer;
 use tokio::sync::watch;
