+++ crates/stages-api/src/pipeline/set.rs
@@ -1,10 +1,10 @@
 use reth_db::database::Database;
 use reth_primitives::stage::StageId;
-use reth_stages_api::Stage;
 use std::{
     collections::HashMap,
     fmt::{Debug, Formatter},
 };
+use crate::Stage;
 
 /// Combines multiple [`Stage`]s into a single unit.
 ///
