+++ crates/stages-api/src/util.rs
